+++ src/main/java/com/hta2405/unite/service/DocService.java
@@ -366,7 +366,7 @@ public class DocService {
         mv.addObject("title", title);
         mv.addObject("message", message);
         mv.addObject("list", docList);
-        mv.setViewName("/doc/list");
+        mv.setViewName("/doc/docList");
         return mv;
     }
