+++ src/main/resources/templates/doc/general_edit.html
@@ -75,7 +75,8 @@
                 </div>
             </div>
         </form>
-        <script th:inline="javascript">
+        <script>
+            const contextPath = /*[[@{/}]]*/ '';
             $(document).ready(function () {
                 $('.summernote').summernote({
                     height: 500, // set editor height
@@ -121,7 +122,7 @@
                         signers: $('input[name="sign[]"]').map(function () {
                             return $(this).val();
                         }).get()
-                    })], { type: "application/json" }));
+                    })], {type: "application/json"}));
 
                     // 파일 추가
                     const fileInput = $('input[name="files"]')[0];
@@ -132,14 +133,14 @@
                     }
 
                     $.ajax({
-                        url: /*[[@{/doc}]]*/'',
+                        url: contextPath + '/api/doc',
                         type: "PATCH",
                         data: formData,
                         processData: false,
                         contentType: false,
                         success: function () {
                             alert("문서 수정이 완료되었습니다.");
-                            window.location.href = /*[[@{/doc/inProgress}]]*/'';
+                            window.location.href = contextPath + '/doc/inProgress';
                         },
                         error: function () {
                             alert("문서 수정 중 오류가 발생했습니다.");
