+++ src/main/resources/templates/doc/sign_edit.html
@@ -43,7 +43,7 @@
 
     <!-- 모달 -->
     <div class="modal fade" id="employeeModal" tabindex="-1" aria-labelledby="employeeModalLabel" aria-hidden="true">
-        <div class="modal-dialog modal-lg">
+        <div class="modal-dialog modal-xl">
             <div class="modal-content">
                 <div class="modal-header">
                     <h5 class="modal-title" id="employeeModalLabel">
