+++ src/main/resources/templates/doc/trip_edit.html
@@ -111,7 +111,8 @@
                 </div>
             </div>
         </form>
-        <script th:inline="javascript">
+        <script>
+            const contextPath = /*[[@{/}]]*/ '';
             $(document).ready(function () {
                 // 제출 버튼 클릭 시 실행
                 $('#submit-button').on('click', function (event) {
@@ -157,7 +158,7 @@
                         signers: $('input[name="sign[]"]').map(function () {
                             return $(this).val();
                         }).get()
-                    })], { type: "application/json" }));
+                    })], {type: "application/json"}));
 
                     // 파일 추가
                     const fileInput = $('input[name="files"]')[0];
@@ -169,14 +170,14 @@
 
                     // AJAX 요청
                     $.ajax({
-                        url: /*[[@{/doc}]]*/'',
+                        url: contextPath + '/doc',
                         type: "PATCH",
                         data: formData,
                         processData: false,
                         contentType: false,
                         success: function () {
                             alert("문서 작성이 완료되었습니다.");
-                            window.location.href = /*[[@{/doc/inProgress}]]*/'';
+                            window.location.href = contextPath + '/doc/inProgress';
                         },
                         error: function () {
                             alert("문서 작성 중 오류가 발생했습니다.");
