+++ src/main/resources/templates/doc/vacation_write.html
@@ -107,7 +107,8 @@
                 </div>
             </div>
         </form>
-        <script th:inline="javascript">
+        <script>
+            const contextPath = /*[[@{/}]]*/ '';
             $(document).ready(function () {
                 // 시작일과 종료일 필드에 change 이벤트 리스너 추가
                 $('input[name="vacation_start"], input[name="vacation_end"]').on('change', function () {
@@ -124,7 +125,7 @@
                         }
 
                         $.ajax({
-                            url: /*[[@{/doc/countVacation}]]*/ '',
+                            url: contextPath + '/api/doc/countVacation',
                             type: "GET",
                             data: {startDate: startDate, endDate: endDate},
                             dataType: "json",
@@ -211,17 +212,16 @@
                         console.log(`${key}:`, value);
                     });
 
-
                     // AJAX 요청
                     $.ajax({
-                        url: /*[[@{/doc/vacation}]]*/'',
+                        url: contextPath + '/api/doc/vacation',
                         type: "POST",
                         data: formData,
                         processData: false,
                         contentType: false,
-                        success: function (response) {
+                        success: function () {
                             alert("문서 작성이 완료되었습니다.");
-                            window.location.href = /*[[@{/doc/inProgress}]]*/'';
+                            window.location.href = contextPath + '/doc/inProgress';
                         },
                         error: function () {
                             alert("문서 작성 중 오류가 발생했습니다.");
